[#15] refactor: Location의 로직을 useLocation 커스텀훅으로 분리
